+++ CHANGELOG.md
@@ -4,3 +4,12 @@
 * Support for local, asset, and network video sources
 * Loop playback with repeatCount parameter
 * Multiple playback callbacks
+
+## 0.0.2
+
+* edit example to use VAP video
+
+
+## 0.0.3
+
+* add FlutterVapController to control video playback